backend/requirements.txt
Django==4.2.17
djangorestframework
pymongo==3.12.3
django-cors-headers
whitenoise
